+++ modules/functionapp/r-function-app.tf
@@ -71,3 +71,10 @@ resource "azurerm_function_app" "function_app" {
 
   tags = merge(var.extra_tags, var.function_app_extra_tags, local.default_tags)
 }
+
+resource "azurerm_app_service_virtual_network_swift_connection" "function_vnet_integration" {
+  count = var.function_app_vnet_integration_enabled ? 1 : 0
+
+  app_service_id = azurerm_function_app.function_app.id
+  subnet_id      = var.function_app_vnet_integration_subnet_id
+}